+++ src/do.c
@@ -733,7 +733,7 @@ int mnum;
         int i;
         for (i = 0; i < 32; i++)
         {
-            unsigned long bit = 1UL << i;
+            uint64_t bit = (uint64_t)1 << i;
             if (((ptr)->mconveys & bit) != 0L)
             {
                 int propnum = conveyed_to_prop(bit);
@@ -3344,7 +3344,7 @@ struct item_description_stats* stats_ptr; /* If non-null, only returns item stat
                     int idx;
                     for (idx = 0; idx < NUM_UNSIGNED_LONG_BITS; idx++)
                     {
-                        unsigned long bit = 1;
+                        uint64_t bit = 1;
                         if (idx > 0)
                             bit = bit << idx;
 
@@ -3428,7 +3428,7 @@ struct item_description_stats* stats_ptr; /* If non-null, only returns item stat
             {
                 for (i = 0; i < MAX_MYTHIC_PREFIX_POWERS; i++)
                 {
-                    unsigned long bit = 1UL << (unsigned long)i;
+                    uint64_t bit = (uint64_t)1 << (uint64_t)i;
                     if (mythic_prefix_powers[i].description && (mythic_prefix_qualities[obj->mythic_prefix].mythic_powers & bit) && mythic_power_applies_to_obj(obj, mythic_prefix_powers[i].power_flags) && strcmp(mythic_prefix_powers[i].description, ""))
                     {
                         powercnt++;
@@ -3452,7 +3452,7 @@ struct item_description_stats* stats_ptr; /* If non-null, only returns item stat
             {
                 for (i = 0; i < MAX_MYTHIC_SUFFIX_POWERS; i++)
                 {
-                    unsigned long bit = 1UL << (unsigned long)i;
+                    uint64_t bit = (uint64_t)1 << (uint64_t)i;
                     if (mythic_suffix_powers[i].description && (mythic_suffix_qualities[obj->mythic_suffix].mythic_powers & bit) && mythic_power_applies_to_obj(obj, mythic_suffix_powers[i].power_flags) && strcmp(mythic_suffix_powers[i].description, ""))
                     {
                         powercnt++;
@@ -3481,8 +3481,8 @@ struct item_description_stats* stats_ptr; /* If non-null, only returns item stat
     boolean show_corpse_hint = FALSE;
     if(stats_known)
     {
-        unsigned long ocflags = obj ? get_obj_oc_flags(obj) : objects[otyp].oc_flags;
-        unsigned long ocflags5 = obj ? get_obj_oc_flags5(obj) : objects[otyp].oc_flags5;
+        uint64_t ocflags = obj ? get_obj_oc_flags(obj) : objects[otyp].oc_flags;
+        uint64_t ocflags5 = obj ? get_obj_oc_flags5(obj) : objects[otyp].oc_flags5;
         /* Item properties */
         if (ocflags & ~(O1_THROWN_WEAPON_ONLY | O1_MELEE_AND_THROWN_WEAPON
             | O1_SPELLTOOL | O1_NON_SPELL_SPELLBOOK | O1_EDIBLE_NONFOOD) 
@@ -3893,9 +3893,9 @@ struct item_description_stats* stats_ptr; /* If non-null, only returns item stat
         }
 
 
-        unsigned long aflags = artilist[obj->oartifact].aflags;
-        unsigned long aflags2 = artilist[obj->oartifact].aflags2;
-        unsigned long mtype = artilist[obj->oartifact].mtype;
+        uint64_t aflags = artilist[obj->oartifact].aflags;
+        uint64_t aflags2 = artilist[obj->oartifact].aflags2;
+        uint64_t mtype = artilist[obj->oartifact].mtype;
         if (aflags & AF_BEHEAD)
         {
             powercnt++;
@@ -4025,7 +4025,7 @@ struct item_description_stats* stats_ptr; /* If non-null, only returns item stat
 
         for (i = 1; i <= 2; i++)
         {
-            unsigned long specialeffect = 0;
+            uint64_t specialeffect = 0;
             Strcpy(endbuf, "");
             if (i == 1)
             {
@@ -4044,7 +4044,7 @@ struct item_description_stats* stats_ptr; /* If non-null, only returns item stat
             int idx;
             for (idx = 0; idx < 32; idx++)
             {
-                unsigned long bit = 1;
+                uint64_t bit = 1;
                 if (idx > 0)
                     bit = bit << idx;
                 propnum = spfx_to_prop(bit);
@@ -4157,7 +4157,7 @@ struct item_description_stats* stats_ptr; /* If non-null, only returns item stat
                 int cnt = 0;
                 for (i = 0; i < NUM_UNSIGNED_LONG_BITS; i++)
                 {
-                    unsigned long bit = 1UL;
+                    uint64_t bit = 1UL;
 
                     if (i > 0)
                         bit = bit << i;
@@ -4169,7 +4169,7 @@ struct item_description_stats* stats_ptr; /* If non-null, only returns item stat
                 int idx = 0;
                 for (i = 0; i < NUM_UNSIGNED_LONG_BITS; i++)
                 {
-                    unsigned long bit = 1UL;
+                    uint64_t bit = 1UL;
 
                     if (i > 0)
                         bit = bit << i;
@@ -8613,7 +8613,7 @@ struct monst* mon;
 struct extended_create_window_info
 extended_create_window_info_from_mon_with_flags(mon, cflags)
 struct monst* mon;
-unsigned long cflags;
+uint64_t cflags;
 {
     struct extended_create_window_info info = { 0 };
     info.monster = mon;
@@ -8853,8 +8853,8 @@ death_hint(VOID_ARGS)
 {
     if ((flags.force_hint || context.game_difficulty <= flags.max_hint_difficulty) && killer.name[0] && killer.hint_idx > 0)
     {
-        unsigned long kbit = 1UL << (killer.hint_idx % 32);
-        unsigned long* hints_given = &u.uhint.kill_hints_given[min(NUM_KILL_HINT_ULONGS - 1, killer.hint_idx / 32)];
+        uint64_t kbit = (uint64_t)1 << (killer.hint_idx % 32);
+        uint64_t* hints_given = &u.uhint.kill_hints_given[min(NUM_KILL_HINT_ULONGS - 1, killer.hint_idx / 32)];
         if (((*hints_given) & kbit) == 0)
         {
             *hints_given |= kbit;
